+++ examples/contract-verification/ethers-v5/scripts/proxy/automatic.ts
@@ -5,7 +5,6 @@ import {
 } from "./deployment";
 
 export async function main() {
-  await sleep(5000);
   // TransparentUpgradeableProxy
   // The plugin will automatically verify the proxy, implementation and all the related contracts on deployment.
   await deployTransparentUpgradeableProxy();
@@ -23,7 +22,3 @@ main().catch((error) => {
   console.error(error);
   process.exitCode = 1;
 });
-
-function sleep(ms: number): Promise<void> {
-  return new Promise(resolve => setTimeout(resolve, ms));
-}