+++ packages/tenderly-core/src/internal/core/types/index.ts
@@ -4,3 +4,6 @@ export * from "./Project";
 export * from "./Requests";
 export * from "./Responses";
 export * from "./TenderlyNetwork";
+export * from "./VerifyContractABIRequest";
+export * from "./VerifyContractABIResponse";
+export * from "./verification-types";