+++ packages/tenderly-core/src/internal/lib/tenderly-lib.ts
@@ -2,3 +2,8 @@ import "../../utils/logger";
 import { TenderlyService } from "../core/services/TenderlyService";
 
 export { TenderlyService };
+export { 
+  VerifyContractABIRequest,
+  VerifyContractABIResponse,
+  VERIFICATION_TYPES,
+} from "../core/types";